+++ uieditor/window-instance.cpp
@@ -216,7 +216,7 @@ int WindowInstance::saveBinToStream(BinStream *stream)
 	int size = 0;
 	unsigned int id;
 	FieldType * ft;
-	int begin_pos;
+	long begin_pos;
 
 	begin_pos = stream->tell();
 	//save stream
@@ -330,9 +330,9 @@ int WindowInstance::saveBinToStream(BinStream *stream)
 	if(count > 0)
 	{
 		//goto offset_props
-		stream->seek(begin_pos+ (int)&(((NCSRM_WINHEADER*)0)->offset_props), StreamStorage::seek_begin);
+		stream->seek(begin_pos+ (long)&(((NCSRM_WINHEADER*)0)->offset_props), StreamStorage::seek_begin);
 		stream->save32(cur_pos1-begin_pos);
-		stream->seek(begin_pos+ (int)&(((NCSRM_WINHEADER*)0)->nr_props),
+		stream->seek(begin_pos+ (long)&(((NCSRM_WINHEADER*)0)->nr_props),
 				StreamStorage::seek_begin);
 		stream->save32(count);
 		size += count * sizeof(Uint32) * 3;
@@ -358,17 +358,17 @@ int WindowInstance::saveBinToStream(BinStream *stream)
 	if(count > 0) //save controls
 	{
 		//goto offset_ctrls
-		stream->seek(begin_pos+ (int)&(((NCSRM_WINHEADER*)0)->offset_ctrls), StreamStorage::seek_begin);
+		stream->seek(begin_pos+ (long)&(((NCSRM_WINHEADER*)0)->offset_ctrls), StreamStorage::seek_begin);
 		stream->save32(cur_pos1-begin_pos);
 		//goto nr_ctrls
-		stream->seek(begin_pos+ (int)&(((NCSRM_WINHEADER*)0)->nr_ctrls), StreamStorage::seek_begin);
+		stream->seek(begin_pos+ (long)&(((NCSRM_WINHEADER*)0)->nr_ctrls), StreamStorage::seek_begin);
 		stream->save32(count);
 	}
 
 	//get size
 	size += ctrls_size;
 	//save size
-	stream->seek(begin_pos + (int)&(((NCSRM_WINHEADER*)0)->size), StreamStorage::seek_begin);
+	stream->seek(begin_pos + (long)&(((NCSRM_WINHEADER*)0)->size), StreamStorage::seek_begin);
 	stream->save32(size);
 
 	stream->seek(cur_pos2, StreamStorage::seek_begin);
